src/requirements.txt
requests==2.5.1
beautifulsoup4==4.8.2
progressbar33==2.4
